Template - In preparation: implementation as h-file
Still builds and works.

Changing the implementation-file into a file that gets included in the
header file is a better practice for templates.

Having them in different files, where the h-file includes the
implementation, is apperemtly the common practice. This is also close to how
C# does it, i.e. implementation and declaration is the same file.

Ref:
http://stackoverflow.com/questions/495021/why-can-templates-only-be-implemented-in-the-header-file