Test <double> added + bugfix. NOTE remaining issue
* va_arg didn't use the template type which is an error
* va_arg fails on template-types of <float>

NOTE: The last point is a remaining issue. Why it's so is worrying and could
mean the template can't be generic enough to be used as template of it's own
type. I.e. the following might not work:

Vector<Vector<int>> 2D_Matrix;

This is kind of important. It's not imperative for the purpose of the
Polygons project, which needs only 1DIM vectors for 3D vertices, but would
disqualify this Vectors class for broader linear algebra context.