Linalgebra operators active again
It's getting very messy. Better would be to have two different classes,
one with linalgebra and one as up until now as base-class or aggregate. That
way the storage-class with it's abilities can be reused for more
vector-based, but not necessarily linalgebra, operations/classes.

Vector-classes of varios specialities could have the same class-name but in
different name-spaces.